Произведён рефакторинг create_category_menu
