+++ menus/create_category_menu/create_category_questions.py
@@ -1,0 +1,9 @@
+from validators.RequiredValidator import RequiredValidator
+
+
+CREATE_CATEGORY_QUESTIONS = [{
+    'type': 'input',
+    'name': 'title',
+    'message': 'Название категории: ',
+    'validate': RequiredValidator
+}]
